requirements.txt
torch==2.0.1 
torchvision==0.15.2 
torchaudio==2.0.2 
pytorch-lightning==2.0.1 
open3d
imageio[ffmpeg]
imageio[pyav]
PyMCubes 
einops 
nerfvis 
open_clip_torch
trimesh 
scipy 
pyglet 
matplotlib 
pyrender 
boto3 
rembg
xformers==0.0.22
fire 
safetensors 
lovely-numpy 
rich
easydict
diffusers
omegaconf
kornia
transformers
tensorboard
webdataset
plyfile
invisible-watermark>=0.2.0
clip @ git+https://github.com/openai/CLIP.git
# openai-clip
gradio==4.18.0
accelerate
git+https://github.com/facebookresearch/segment-anything.git
pygltflib==1.16.1
MiniballCpp==0.2.3
xatlas
fast-simplification
git+https://github.com/NVlabs/nvdiffrast.git@c5caf7bdb8a2448acc491a9faa47753972edd380
fpsample
